refactor executor
