+++ test/e2e_operator/common/helm.go
@@ -17,12 +17,9 @@ limitations under the License.
 package common
 
 import (
-	"bytes"
 	"fmt"
-	"os/exec"
-	"strings"
 
-	e2elog "k8s.io/kubernetes/test/e2e/framework/log"
+	"github.com/dell/csi-baremetal/pkg/base/command"
 )
 
 type HelmExecutor interface {
@@ -33,6 +30,7 @@ type HelmExecutor interface {
 // CmdHelmExecutor is HelmExecutor implementation using os/exec.Cmd
 type CmdHelmExecutor struct {
 	kubeconfig string
+	executor   command.CmdExecutor
 }
 
 // HelmChart stores info about chart in filesystem
@@ -48,7 +46,9 @@ func (c *CmdHelmExecutor) InstallRelease(ch *HelmChart, args string) error {
 	cmdStr := fmt.Sprintf("helm install %s %s --kubeconfig %s -n %s "+args,
 		ch.name, ch.path, c.kubeconfig, ch.namespace)
 
-	return execCmdStr(cmdStr)
+	_, _, err := c.executor.RunCmd(cmdStr)
+
+	return err
 }
 
 // DeleteRelease call "helm delete" for chart
@@ -58,26 +58,7 @@ func (c *CmdHelmExecutor) DeleteRelease(ch *HelmChart) error {
 		"-n %s "+"%s",
 		c.kubeconfig, ch.namespace, ch.name)
 
-	return execCmdStr(cmdStr)
-}
-
-func execCmdStr(cmdStr string) error {
-	cmdStrSplit := strings.Split(cmdStr, " ")
-	return execCmdObj(exec.Command(cmdStrSplit[0], cmdStrSplit[1:]...))
-}
-
-func execCmdObj(cmd *exec.Cmd) error {
-	var stdout, stderr bytes.Buffer
-
-	e2elog.Logf("Exec: %s", cmd.Args)
-
-	cmd.Stdout = &stdout
-	cmd.Stderr = &stderr
-
-	err := cmd.Run()
-
-	e2elog.Logf("Stdout: %s", stdout.String())
-	e2elog.Logf("Stderr: %s", stderr.String())
+	_, _, err := c.executor.RunCmd(cmdStr)
 
 	return err
 }
