+++ test/e2e_operator/scenarios/different-sc-test.go
@@ -94,7 +94,7 @@ func differentSCTypesTest(driver testsuites.TestDriver) {
 		pod = startAndWaitForPodWithPVCRunning(f, ns, pvcs)
 	})
 
-	/*ginkgo.It("should create Pod with PVC with ANY type", func() {
+	ginkgo.It("should create Pod with PVC with ANY type", func() {
 		scType := "ANY"
 		init(scType)
 		defer cleanup()
@@ -135,7 +135,7 @@ func differentSCTypesTest(driver testsuites.TestDriver) {
 		pvcs = []*corev1.PersistentVolumeClaim{createBlockPVC(
 			f, 1, driver.(testsuites.DynamicPVTestDriver).GetClaimSize(), k8sSC.Name, ns)}
 		pod = startAndWaitForPodWithPVCRunning(f, ns, pvcs)
-	})*/
+	})
 }
 
 func createBlockPVC(f *framework.Framework, numberOfPVC int, size string, scName string, ns string) *corev1.PersistentVolumeClaim {
